almost there, still requires fixing of margins or padding in the template of the photo-comments route (oops forgot some)
